# swagger.yaml
openapi: 3.0.0
info:
  title: Reactors API Documentation
  version: 1.0.0
  description: API documentation for the Reactors Book Tracker project
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in HttpOnly cookie
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Something went wrong
        error:
          type: string
          nullable: true
...
    Book:
      type: object
      properties:
        _id:
          type: string
          example: 67411c5c0a8c5d3b85e0da11
        title:
          type: string
          example: The Pragmatic Programmer
        author:
          type: string
          example: Andrew Hunt
        genre:
          type: string
          nullable: true
          example: Programming
        description:
          type: string
          nullable: true
          example: A practical guide to software craftsmanship.
        coverImage:
          type: string
          nullable: true
          example: https://example.com/cover.jpg
        createdBy:
          type: string
          example: 691ce9f36a70ea178440d121
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateBookRequest:
      type: object
      required:
        - title
        - author
      properties:
        title:
          type: string
          example: The Pragmatic Programmer
        author:
          type: string
          example: Andrew Hunt
        genre:
          type: string
          example: Programming
        description:
          type: string
          example: A practical guide to software craftsmanship.
        coverImage:
          type: string
          example: https://example.com/cover.jpg
    UpdateBookRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        description:
          type: string
        coverImage:
          type: string
    UserBook:
      type: object
      properties:
        _id:
          type: string
        user:
          type: string
          example: 691ce9f36a70ea178440d121
        book:
          oneOf:
            - type: string
            - $ref: "#/components/schemas/Book"
        status:
          type: string
          enum: [wishlist, reading, finished]
          example: reading
        progress:
          type: integer
          minimum: 0
          maximum: 100
          example: 40
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
          example: 4
        notes:
          type: string
          nullable: true
          example: Really liked the examples.
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserBookRequest:
      type: object
      required:
        - bookId
      properties:
        bookId:
          type: string
          example: 67411c5c0a8c5d3b85e0da11
        status:
          type: string
          enum: [wishlist, reading, finished]
          example: wishlist
        progress:
          type: integer
          minimum: 0
          maximum: 100
          example: 0
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        notes:
          type: string
          nullable: true
          example: Want to read this next.
    UpdateUserBookRequest:
      type: object
      properties:
        status:
          type: string
          enum: [wishlist, reading, finished]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        rating:
          type: integer
          minimum: 1
          maximum: 5
        notes:
          type: string
    UserBookSummary:
      type: object
      properties:
        wishlist:
          type: integer
          example: 2
        reading:
          type: integer
          example: 1
        finished:
          type: integer
          example: 3
        avgRating:
          type: number
          format: float
          nullable: true
          example: 4.3
        totalProgress:
          type: integer
          example: 240
        lastUpdated:
          type: string
          format: date-time
          nullable: true
...
paths:
...
  /api/books:
    post:
      security:
        - cookieAuth: []
      tags: [Books]
      summary: Create a new book
      description: Creates a new book and assigns it to the authenticated user as creator.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookRequest"
      responses:
        "201":
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: Validation error (missing/invalid title or author, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
        "500":
          description: Server error
    get:
      tags: [Books]
      summary: Get all books
      description: Returns a list of all books in the system.
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
        "500":
          description: Server error
...
  /api/books/{id}:
    get:
      tags: [Books]
      summary: Get a single book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Book found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
    put:
      security:
        - cookieAuth: []
      tags: [Books]
      summary: Update a book (creator only)
      description: Only the user who created the book can update it.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookRequest"
      responses:
        "200":
          description: Book updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: Validation error (e.g., title too short)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed (not the creator)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Book not found
        "401":
          description: Not authenticated
    delete:
      security:
        - cookieAuth: []
      tags: [Books]
      summary: Delete a book (creator only)
      description: Only the user who created the book can delete it.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Book deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Book deleted successfully
        "403":
          description: Not allowed (not the creator)
        "404":
          description: Book not found
        "401":
          description: Not authenticated
        "500":
          description: Server error
  /api/userbooks:
    post:
      security:
        - cookieAuth: []
      tags: [UserBooks]
      summary: Add a book to the user's reading list
      description: Creates a UserBook entry for the logged-in user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBookRequest"
      responses:
        "201":
          description: Entry created in reading list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBook"
        "400":
          description: Validation error or book already in list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
        "500":
          description: Server error
    get:
      security:
        - cookieAuth: []
      tags: [UserBooks]
      summary: Get the logged-in user's reading list
      description: Returns all UserBook entries for the logged-in user, populated with book details.
      responses:
        "200":
          description: Reading list for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserBook"
        "401":
          description: Not authenticated
        "500":
          description: Server error
  /api/userbooks/summary:
    get:
      security:
        - cookieAuth: []
      tags: [UserBooks]
      summary: Get summary statistics of the user's reading list
      description: Returns counts for wishlist/reading/finished, average rating, total progress, and lastUpdated.
      responses:
        "200":
          description: Summary of the user's reading list
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Reading list summary retrieved successfully
                  summary:
                    $ref: "#/components/schemas/UserBookSummary"
        "401":
          description: Not authenticated
        "500":
          description: Server error
  /api/userbooks/{id}:
    put:
      security:
        - cookieAuth: []
      tags: [UserBooks]
      summary: Update a reading list entry
      description: Only the owner of the entry can update it.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBookRequest"
      responses:
        "200":
          description: UserBook entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBook"
        "400":
          description: Validation error or attempt to change book field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed (not the owner)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Entry not found
        "401":
          description: Not authenticated
        "500":
          description: Server error
    delete:
      security:
        - cookieAuth: []
      tags: [UserBooks]
      summary: Delete a reading list entry
      description: Only the owner of the entry can delete it.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entry deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Entry deleted
        "403":
          description: Not allowed (not the owner)
        "404":
          description: Entry not found
        "401":
          description: Not authenticated
        "500":
          description: Server error